update visualization tokoens to be prefixed with dataviz
