+++ .changeset/wicked-seahorses-flash.md
@@ -1,0 +1,5 @@
+---
+"@hopper-ui/tokens": major
+---
+
+prefixed data visualization tokens with "dataviz"
