+++ packages/tokens/src/tokens/semantic/dark/dataViz.tokens.json
@@ -1,22 +1,24 @@
 {
-  "unavailable": {
-    "$type": "color",
-    "$value": "{rock.600}"
-  },
-  "unavailable-weak": {
-    "$type": "color",
-    "$value": "{rock.400}"
-  },
-  "unavailable-strong": {
-    "$type": "color",
-    "$value": "{rock.700}"
-  },
-  "text-onlight": {
-    "$type": "color",
-    "$value": "{rock.800}"
-  },
-  "text-ondark": {
-    "$type": "color",
-    "$value": "{samoyed}"
-  }
+    "dataviz": {
+        "unavailable": {
+            "$type": "color",
+            "$value": "{rock.600}"
+        },
+        "unavailable-weak": {
+            "$type": "color",
+            "$value": "{rock.400}"
+        },
+        "unavailable-strong": {
+            "$type": "color",
+            "$value": "{rock.700}"
+        },
+        "text-onlight": {
+            "$type": "color",
+            "$value": "{rock.800}"
+        },
+        "text-ondark": {
+            "$type": "color",
+            "$value": "{samoyed}"
+        }
+    }
 }
